fix a few more pointer truncations in windows dialog stuff a few printf arguments need a cast to int

+++ src/drivers/win/texthook.cpp
@@ -1494,7 +1494,7 @@ int excitecojp() {
 	memset( buffer, 0, 2048 );
 	
 	//form the http request
-	sprintf( buffer, "POST /world/english/ HTTP/1.1\r\nHost: www.excite.co.jp\r\nContent-Type: application/x-www-form-urlencoded\r\nContent-Length: %d\r\n\r\nbefore=%s&wb_lp=JAEN&start=%%96%%7C+%%96%%F3", strlen( urlEscapedText ) + 38, urlEscapedText );
+	sprintf( buffer, "POST /world/english/ HTTP/1.1\r\nHost: www.excite.co.jp\r\nContent-Type: application/x-www-form-urlencoded\r\nContent-Length: %d\r\n\r\nbefore=%s&wb_lp=JAEN&start=%%96%%7C+%%96%%F3", (int)strlen( urlEscapedText ) + 38, urlEscapedText );
 		
 	//this was needed before in order to make it work, keeping it here in case it starts acting up again
 	//memset( textToTrans, 0, 1024 );
